requirements.txt
# --- ANALYZER REQUIREMENTS ---
easyocr
matplotlib
numpy
paddleocr
# paddlepaddle
pandas
scenedetect
scikit-image
opencv-python-headless

# --- CHATBOT REQUIREMENTS ---
python-dotenv
requests
groq
# --- EDITOR REQUIREMENTS ---
pillow
shotstack-sdk

# --- VIDEO DOWNLOAD & CLIPPING ---
yt-dlp

# --- UTILITIES ---
fastapi
google-api-python-client
google-auth-oauthlib
python-multipart
typing-extensions
uvicorn[standard]
